+++ src/components/UI/navbar/navbar.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import {Link} from "react-router-dom";
+
+const Navbar = () => {
+    return (
+        <div className="navbar">
+            <div className="navbar__links">
+                <Link to="/about">About</Link>
+                <Link to="/posts">Posts</Link>
+            </div>
+        </div>
+    );
+};
+
+export default Navbar;
