+++ src/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const About = () => {
+    return (
+        <div>
+           <h1> This app create with course of React js</h1>
+        </div>
+    );
+};
+
+export default About;
