+++ src/pages/Error.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Error = () => {
+    return (
+        <div>
+            <h1 style={{color: 'red'}}>
+                Error: Page not found!
+            </h1>
+        </div>
+    );
+};
+
+export default Error;
